requirements.txt
# Core dependencies
langchain==0.1.13
openai==1.12.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
chromadb==0.4.22
numpy==1.26.4

# Procesamiento de datos
pydantic==2.6.4
pandas==2.2.1
tqdm==4.66.2

# Interfaz
curses-menu==1.0.0
rich==13.7.1

# Machine Learning
torch==2.2.1
transformers==4.38.2
sentence-transformers==2.5.1
peft==0.9.0
trl==0.7.11